Customize port
